init commit, remove switch case, move files to own folder
